[1760]: Move the tree command to its own file so that its easier to snag and move to other versions if needed. Also output an error if no commands are found.
